text-in-image/requirements.txt
flask~=3.0.0
python-dotenv~=1.0.0
openai
Pillow~=10.2.0
numpy~=1.26.0
rembg~=2.0.0
onnxruntime